Add postrges, Dockerfile and docker-compose

+++ src/notes/service.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException
 from src.core import database as db
 from src.utils.web import raise_404_if_none
 from src.core.repository import make_sqlalchemy_repo
@@ -14,7 +15,7 @@ repo = make_sqlalchemy_repo()
 async def get_filtered_notes(
     filters: dict, limit: int = None, scheme: BaseNoteModel = NoteOut,
 ) -> list[BaseNoteModel]:
-    notes = await repo.filter_all(db.NoteModel, filters, limit)
+    notes = await repo.filter_by(db.NoteModel, filters, limit)
     return [scheme(**note.__dict__) for note in notes]
 
 
@@ -33,25 +34,36 @@ async def create_note(
 
 
 async def get_note(
-    note_id: int, scheme: BaseNoteModel = NoteOut
+    note_id: int, user: db.UserModel, scheme: BaseNoteModel = NoteOut
 ) -> BaseNoteOutModel:
     note = await repo.get(db.NoteModel, note_id)
     await raise_404_if_none(
         note, message=f'Note with id \'{note_id}\' not found'
     )
+    if note.author_id != user.id:
+        raise HTTPException(403, detail='Permission denied')
     return scheme(**note.__dict__)
 
 
 async def update_note(
-    data: dict, note_id: int, scheme: NoteOut = NoteOut
+    data: dict, note_id: int, user: db.UserModel, scheme: NoteOut = NoteOut
 ) -> NoteOut:
+    note = await repo.get(db.NoteModel, note_id)
+    await raise_404_if_none(
+        note, message=f'Note with id \'{note_id}\' not found'
+    )
+    if note.author_id != user.id:
+        raise HTTPException(403, detail='Permission denied')
     updated_note = await repo.update(db.NoteModel, pk=note_id, values=data)
     return scheme(**updated_note.__dict__)
 
 
-async def delete_note(note_id: int) -> int:
-    deleted_note_id = await repo.delete(db.NoteModel, pk=note_id)
+async def delete_note(note_id: int, user: db.UserModel) -> int:
+    note = await repo.get(db.NoteModel, note_id)
     await raise_404_if_none(
-        deleted_note_id, message=f'Note with id \'{note_id}\' not found'
+        note, message=f'Note with id \'{note_id}\' not found'
     )
+    if note.author_id != user.id:
+        raise HTTPException(403, detail='Permission denied')
+    deleted_note_id = await repo.delete(db.NoteModel, pk=note_id)
     return deleted_note_id
